pyproject.toml
[project]
name = "data-forge"
version = "0.1.0"
description = "Schema-aware synthetic data platform for realistic, relational, time-aware test data"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Data Forge" }]
keywords = ["synthetic-data", "testing", "schema", "etl", "data-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlmodel>=0.0.16",
    "duckdb>=1.1.0",
    "psycopg[binary]>=3.0.0",
    "snowflake-connector-python>=3.0.0",
    "google-cloud-bigquery>=3.0.0",
    "polars>=1.0.0",
    "pyarrow>=18.0.0",
    "faker>=33.0.0",
    "mimesis>=7.0.0",
    "pyyaml>=6.0.2",
    "jsonschema>=4.0.0",
    "jinja2>=3.1.4",
    "great-expectations>=0.18.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "hypothesis>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
sdv = ["sdv>=1.11.0"]

[project.scripts]
data-forge = "data_forge.cli:app"
data-forge-api = "data_forge.api.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/data_forge"]

[tool.hatch.build.targets.sdist]
packages = ["src/data_forge"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
